[TASK] - Move movie genres to design module
