+++ features/mpmoviedetails/src/main/res/layout/list_item_movie_detail_genre.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="viewState"
-            type="com.jpp.mpmoviedetails.MovieGenreItem" />
+            type="com.jpp.mpdesign.mapped.MovieGenreItem" />
     </data>
