+++ mpdesign/src/main/res/drawable-night/ic_adventure.xml
