+++ mpdesign/src/main/res/drawable-night/ic_animation.xml
