+++ mpdesign/src/main/res/drawable-night/ic_comedy.xml
