+++ mpdesign/src/main/res/drawable-night/ic_documentary.xml
