+++ mpdesign/src/main/res/drawable-night/ic_drama.xml
