+++ mpdesign/src/main/res/drawable-night/ic_generic.xml
