+++ mpdesign/src/main/res/drawable-night/ic_mystery.xml
