+++ mpdesign/src/main/res/drawable-night/ic_science_ficcion.xml
