+++ mpdesign/src/main/res/drawable-night/ic_thriller.xml
