+++ mpdesign/src/main/res/drawable-night/ic_war.xml
