+++ mpdesign/src/main/res/drawable-night/ic_western.xml
