+++ mpdesign/src/main/res/drawable/ic_adventure.xml
