+++ mpdesign/src/main/res/drawable/ic_animation.xml
