+++ mpdesign/src/main/res/drawable/ic_crime.xml
