+++ mpdesign/src/main/res/drawable/ic_drama.xml
