+++ mpdesign/src/main/res/drawable/ic_generic.xml
