+++ mpdesign/src/main/res/drawable/ic_science_ficcion.xml
