+++ mpdesign/src/main/res/drawable/ic_war.xml
