+++ mpdesign/src/main/res/drawable/ic_western.xml
