+++ mpdesign/src/main/res/values-es/strings.xml
@@ -10,4 +10,26 @@
 
     <string name="login_generic">Iniciar Sesión</string>
     <string name="about_generic">Acerca de</string>
+
+    <!-- Genres -->
+    <string name="action_genre">Acción</string>
+    <string name="adventure_genre">Aventura</string>
+    <string name="animation_genre">Animación</string>
+    <string name="comedy_genre">Comedia</string>
+    <string name="crime_genre">Crimen</string>
+    <string name="documentary_genre">Documental</string>
+    <string name="drama_genre">Drama</string>
+    <string name="familiy_genre">Familia</string>
+    <string name="fantasy_genre">Fantasia</string>
+    <string name="history_genre">Historia</string>
+    <string name="horror_genre">Terror</string>
+    <string name="music_genre">Música</string>
+    <string name="mystery_genre">Misterio</string>
+    <string name="sci_fi_genre">Ciencia Ficción</string>
+    <string name="tv_movie_genre">Película de Tv</string>
+    <string name="thriller_genre">Thriller</string>
+    <string name="war_genre">Guerra</string>
+    <string name="western_genre">Western</string>
+    <string name="generic_genre">Genérico</string>
+    <!-- Genres -->
 </resources>
