+++ mpdesign/src/main/res/values/strings.xml
@@ -10,4 +10,28 @@
 
     <string name="login_generic">Login</string>
     <string name="about_generic">About</string>
+
+    <!-- Genres -->
+    <string name="action_genre">Action</string>
+    <string name="adventure_genre">Adventure</string>
+    <string name="animation_genre">Animation</string>
+    <string name="comedy_genre">Comedy</string>
+    <string name="crime_genre">Crime</string>
+    <string name="documentary_genre">Documentary</string>
+    <string name="drama_genre">Drama</string>
+    <string name="familiy_genre">Family</string>
+    <string name="fantasy_genre">Fantasy</string>
+    <string name="history_genre">History</string>
+    <string name="horror_genre">Horror</string>
+    <string name="music_genre">Music</string>
+    <string name="mystery_genre">Mystery</string>
+    <string name="sci_fi_genre">SciFi</string>
+    <string name="tv_movie_genre">Tv Movie</string>
+    <string name="thriller_genre">Thriller</string>
+    <string name="war_genre">War</string>
+    <string name="western_genre">Western</string>
+    <string name="generic_genre">Generic</string>
+    <string name="movie_credits_title">Credits</string>
+    <!-- Genres -->
+
 </resources>
